added logic and setup a user to execute the mumbler_app, which calls the three nodes to get words back

+++ mumbler/process_ngram_zip.py
@@ -1,9 +1,11 @@
+#!/usr/local/bin/python2.7
 from configparser import ConfigParser
 import glob
 import io
 from io import TextIOWrapper
 import json
 import os
+from os.path import basename
 import re
 import sys
 import time
@@ -16,42 +18,71 @@ JSON_DIR = DATA_DIR+"/json"
 PARSE_CSV_FILE_NAME_PATTERN = DATA_DIR+"/{}.{}.parse.csv"
 IGNORED_LINES_PATTERN = DATA_DIR+"/ignored.lines.{}.csv"
 
+save_ignored_lines = False
+compress_result_files = False
+
 # The following regex template matches a single starting character, specified per node, followed by any characters that would match for any node,
 # then optionally matches a space and any set of characters that would could be matched by any node, followed by year, match_count, page_count and
 # volume_count
-NGRAM_REGEX = "(?P<word1>{}+[a-zA-Z.'\"]*)(?P<word2>\ [a-zA-Z.'\"]*)?\t(?P<year>\d+)\t(?P<match_count>\d+)\t(?P<page_count>\d+)\t(?P<volume_count>\d+)"
+#NGRAM_REGEX_FORMAT = "(?P<word1>{}+[a-zA-Z]*)(?P<word2>[ ][a-zA-Z]*)?\t(?P<year>d+)\t(?P<match_count>d+)\t(?P<page_count>d+)\t(?P<volume_count>d+)"
+NGRAM_REGEX_FORMAT = "(?P<word1>[a-zA-Z\'\"\.]+[a-zA-Z\'\"\.]*) ?(?P<word2>[a-zA-Z\'\"\.]*)?\t(?P<year>\d+)\t(?P<match_count>\d+)\t(?P<page_count>\d+)\t(?P<volume_count>\d+)"
 
-#node_regex = None
-#config = configparser.RawConfigParser()
-#letters_words_counts = {}
+
+def only_ascii(test_string):
+    return re.match('^[\x00-\x7F]+$', test_string)
+
+
+def write_ignored_line(line, node_id):
+    global ignored_lines_file
+
+    if not ignored_lines_file:
+        ignored_lines_file = io.open(IGNORED_LINES_PATTERN.format(node_id), mode="a+", encoding="utf-8")
+    ignored_lines_file.write(line)
+
 
 def process_zip_file(zip_file_path, letters_words_counts):
     global node_regex
     global ignored_lines_file
 
-    with zipfile.ZipFile(zip_file_path, "r") as zip_file:
+    # The following regex will be used for matching all lines with ngrams in English including optional . ' and "
+    ngram_regex = re.compile(NGRAM_REGEX_FORMAT)
+
+    with zipfile.ZipFile(zip_file_path, "r", allowZip64=True) as zip_file:
         for csv_filename in zip_file.namelist():
-            with zip_file.open(csv_filename, "r") as csv_file:
-                for line in TextIOWrapper(csv_file, encoding = "utf-8", errors="backslashreplace", newline=None):
+            with zip_file.open(csv_filename) as csv_file:
+                #for line in TextIOWrapper(csv_file, encoding = "ascii", errors="backslashreplace", newline=None):
+                #for line in TextIOWrapper(csv_file, encoding = None, errors="strict", newline=None):
+                for line in csv_file:
 
-                    reg_match = node_regex.match(line)
-                    if reg_match:
-                        word1 = reg_match.group('word1')
-                        word2 = reg_match.group('word2')
-                        year = reg_match.group('year')
-                        match_count = reg_match.group('match_count')
-                        page_count = reg_match.group('page_count')
-                        volume_count = reg_match.group('volume_count')
-                        # print("ngram: \"{}\", year: \"{}\", match_count: \"{}\", page_count: \"{}\", volume_count: \"{}\"".format(ngram, year, match_count, page_count, volume_count))
+                    # Discard non-ascii lines, for simplicity
+                    if only_ascii(line):
+                        #line = str(line)
+                        reg_match = ngram_regex.match(line)
 
-                        process_ngram(word1, word2, int(match_count), letters_words_counts)
-                    else:
-                        write_other_node_file(line)
+                        # Also determine if the line should be parsed by this node
+                        reg_match = node_regex.match(line[0])
+
+                        if reg_match:
+                            word1 = reg_match.group('word1')
+                            word2 = reg_match.group('word2')
+
+                            # remove the leading space included by the regex
+                            word2 = word2[1:]
+
+                            year = reg_match.group('year')
+                            match_count = reg_match.group('match_count')
+                            page_count = reg_match.group('page_count')
+                            volume_count = reg_match.group('volume_count')
+
+                            process_ngram(word1, word2, int(match_count), letters_words_counts)
+                        else:
+                            write_other_node_file(line)
 
 def close_files():
 
     PARSE_ZIP_FILE_NAME_PATTERN = DATA_DIR+"/{}.{}.parse.csv.zip"
 
+    global compress_result_files
     global ignored_lines_file
     global other_nodes_csv_files
 
@@ -64,26 +95,27 @@ def close_files():
             csv_file.flush()
             csv_file.close()
 
-    # iterate over all the parse files for other nodes, put them into zips and delete the original files
-    for path in glob.glob(DATA_DIR+"/gpfs?."+node_id+".parse.csv"):
-        zip_file_path = path + ".zip"
+    if compress_result_files:
+        # iterate over all the parse files for other nodes, put them into zips and delete the original files
+        for path in glob.glob(DATA_DIR+"/gpfs?."+node_id+".parse.csv"):
+            zip_file_path = path + ".zip"
 
-        specific_path = path + "." + str(time.time())
-        os.rename(path, specific_path)
+            specific_path = path + "." + str(time.time())
+            os.rename(path, specific_path)
 
-        with zipfile.ZipFile(zip_file_path, "a") as zip_file:
-            zip_file.write(specific_path)
+            with zipfile.ZipFile(zip_file_path, "a", allowZip64=True) as zip_file:
+                zip_file.write(specific_path, basename(specific_path))
 
-        os.remove(specific_path)
+            os.remove(specific_path)
 
-    # iterate over all the JSON letter files, put them into zips and delete the original files
-    zip_file_path = str(JSON_DIR+"/letter_trees." + node_id + "." + str(time.time()) + ".zip")
-    for path in glob.glob(JSON_DIR+"/*_tree."+node_id+".*.json"):
-        zip_file_path = path+".zip"
-        with zipfile.ZipFile(zip_file_path, "w") as zip_file:
-            zip_file.write(path)
+        # iterate over all the JSON letter files, put them into zips and delete the original files
+        for path in glob.glob(JSON_DIR+"/*_tree."+node_id+".*.json"):
 
-        os.remove(path)
+            zip_file_path = path+".zip"
+            with zipfile.ZipFile(zip_file_path, "w", allowZip64=True) as zip_file:
+                zip_file.write(path, basename(path))
+
+            os.remove(path)
 
     # zip the ignored lines file and delete the original
     ignored_lines_filename = IGNORED_LINES_PATTERN.format(node_id)
@@ -95,15 +127,15 @@ def close_files():
         ignored_specific_filename = ignored_lines_filename + "." + str(time.time())
         os.rename(ignored_lines_filename, ignored_specific_filename)
 
-        with zipfile.ZipFile(ignored_zip, "a") as zip_file:
-            zip_file.write(ignored_specific_filename)
+        with zipfile.ZipFile(ignored_zip, "a", allowZip64=True) as zip_file:
+            zip_file.write(ignored_specific_filename, basename(ignored_specific_filename))
 
         os.remove(ignored_specific_filename)
 
 
 def write_other_node_file(line):
 
-    global ignored_lines_file
+    global save_ignored_lines
     global other_nodes_csv_files
     global other_nodes_zip_files
 
@@ -111,7 +143,7 @@ def write_other_node_file(line):
     ignored_lines_filename = IGNORED_LINES_PATTERN.format(node_id)
 
     for some_node_id, some_regex in other_nodes_regex.iteritems():
-        other_regex = re.compile(NGRAM_REGEX.format(some_regex, some_regex))
+        other_regex = re.compile(some_regex)
         if other_regex.match(line):
             matched = True
 
@@ -119,15 +151,13 @@ def write_other_node_file(line):
             other_node_csv_file = other_nodes_csv_files.get(some_node_id)
 
             if not other_node_csv_file:
-                other_node_csv_file = io.open(other_node_csv_file_path, mode="a+", encoding="utf-8")
+                other_node_csv_file = io.open(other_node_csv_file_path, mode="a+", encoding="ascii")
                 other_nodes_csv_files[some_node_id] = other_node_csv_file
 
             other_node_csv_file.write(line)
 
-    if not matched:
-        if not ignored_lines_file:
-            ignored_lines_file = io.open(IGNORED_LINES_PATTERN.format(node_id), mode="a+", encoding="utf-8")
-        ignored_lines_file.write(line)
+    if (not matched) & save_ignored_lines:
+        write_ignored_line(line, node_id)
 
 def process_ngram(first_word, second_word, match_count, letters_words_counts):
 
@@ -179,7 +209,7 @@ def split_by_letter(letter_words_counts):
 
     letter_trees = {}
 
-    for word, counts in letter_words_counts.iteritems():
+    for word, counts in letter_words_counts["next"].iteritems():
 
         letter_key = word[0]
 
@@ -195,10 +225,12 @@ def split_by_letter(letter_words_counts):
         letter_dict[word] = counts
 
     for letter, tree in letter_trees.iteritems():
-        filepath = JSON_DIR+"/"+letter+"_tree."+node_id+"."+str(time.time())+".json"
-        with io.open(filepath, mode="w", encoding="ascii") as dict_file:
+        file_path = JSON_DIR+"/"+letter+"_tree."+node_id+"."+str(time.time())+".json"
+        with io.open(file_path, mode="w", encoding="ascii") as dict_file:
             json_string = json.dumps(tree)
-            dict_file.write(json_string)
+            dict_file.write(unicode(json_string))
+            dict_file.flush()
+            dict_file.close()
 
 def read_config():
 
@@ -206,7 +238,7 @@ def read_config():
     global letters_words_counts
     global node_regex
     global other_nodes_regex
-    global ignored_lines_filenamea
+    global ignored_lines_filename
 
     config = ConfigParser()
     letters_words_counts = {}
@@ -214,7 +246,7 @@ def read_config():
 
     if os.path.isfile("mumbler.parse.cfg"):
         try:
-            config.read("mumbler.parse.cfg",encoding="utf-8")
+            config.read("mumbler.parse.cfg")
         except Exception as e:
             print("Couldn't read config file: "+str(e))
     else:
@@ -234,7 +266,7 @@ def read_config():
         print("Couldn't read config section and variable: "+str(e))
 
     print "node_regex_conf: "+ node_regex_conf
-    node_regex = re.compile(NGRAM_REGEX.format(node_regex_conf))
+    node_regex = re.compile(node_regex_conf)
 
 if __name__ == "__main__":
 
@@ -272,7 +304,6 @@ if __name__ == "__main__":
     ignored_lines_filename = IGNORED_LINES_PATTERN.format(node_id)
 
 
-    print("Command line: ", str(sys.argv))
     if len(sys.argv) != 2:
         print("\nUsage: python process_ngram_zip <zip file path>")
     else:
